[Refactor] BookMatchKit Entity 명칭 일관되게 변경 (#66)
