+++ BookKitty/BookKitty/BookMatchKit/Sources/BookMatchAPI/Protocols/AIRecommendable.swift
@@ -3,8 +3,8 @@ import RxSwift
 
 public protocol AIRecommendable {
     func getBookRecommendation(question: String, ownedBooks: [OwnedBook])
-        -> Single<AiRecommendationForQuestion>
-    func getBookRecommendation(ownedBooks: [OwnedBook]) -> Single<AiRecommendationFromOwnedBooks>
+        -> Single<RecommendationForQuestion>
+    func getBookRecommendation(ownedBooks: [OwnedBook]) -> Single<RecommendationFromOwned>
     func getAdditionalBook(question: String, previousBooks: [RawBook]) -> Single<RawBook>
     func getDescription(question: String, books: [RawBook]) -> Single<String>
 }
