+++ BookKitty/BookKitty/BookMatchKit/Sources/BookMatchCore/Models/BookItem.swift
@@ -41,40 +41,3 @@ public struct BookItem: Codable, Identifiable, Hashable {
         self.pubdate = pubdate
     }
 }
-
-/// 기본적인 도서 정보(제목, 저자)를 나타내는 구조체입니다.
-/// GPT 모델과의 통신에 사용됩니다.
-public struct RawBook: Codable, Hashable {
-    // MARK: - Properties
-
-    // MARK: - Public
-
-    public let title: String
-    public let author: String
-
-    // MARK: - Lifecycle
-
-    public init(title: String, author: String) {
-        self.title = title
-        self.author = author
-    }
-}
-
-/// 사용자가 보유한 도서 정보를 나타내는 구조체입니다.
-public struct OwnedBook: Codable, Identifiable, Hashable {
-    // MARK: - Properties
-
-    // MARK: - Public
-
-    public let id: String // ISBN
-    public let title: String
-    public let author: String
-
-    // MARK: - Lifecycle
-
-    public init(id: String, title: String, author: String) {
-        self.id = id
-        self.title = title
-        self.author = author
-    }
-}
