+++ BookKitty/BookKitty/BookMatchKit/Sources/BookMatchCore/Models/RecommendationForQuestion.swift
@@ -1,5 +1,5 @@
 /// `사용자 질문 기반 도서 추천` 결과를 담는 구조체입니다.
-public struct AiRecommendationForQuestion: Codable {
+public struct RecommendationForQuestion: Codable {
     // MARK: - Properties
 
     public let ownedBooks: [OwnedBook]
@@ -13,15 +13,3 @@ public struct AiRecommendationForQuestion: Codable {
     }
 }
 
-/// `보유 도서 기반 도서 추천` 결과를 담는 구조체입니다.
-public struct AiRecommendationFromOwnedBooks: Codable {
-    // MARK: - Properties
-
-    public let books: [RawBook]
-
-    // MARK: - Lifecycle
-
-    public init(books: [RawBook]) {
-        self.books = books
-    }
-}
